= SipUserAgent =  * doxygen 을 위한 주석을 추가함.
= TestSipParser =
 * 64bit 에서 warning 없이 컴파일되도록 수정함.

+++ TestSipParser/TestSdp.cpp
@@ -21,7 +21,7 @@
 static bool Test( const char * pszText, const char * pszResult )
 {
 	CSdpMessage	clsSdp;
-	int		iPos, iLen, iTextLen = strlen(pszText);
+	int		iPos, iLen, iTextLen = (int)strlen(pszText);
 	char	szResult[512];
 
 	memset( szResult, 0, sizeof(szResult) );
